[optimize] 优化圆点显示逻辑

+++ module_mine/src/main/java/com/mredrock/cyxbs/mine/page/feedback/network/bean/history/HistoryFeedback.kt
@@ -31,7 +31,7 @@ data class Feedback(
     @SerializedName("DeletedAt")
     val deletedAt: Any,
     @SerializedName("ID")
-    val iD: Int,
+    val iD: Long,
     @SerializedName("name")
     val name: String,
     @SerializedName("product_id")
